Fix datatype of boxes

+++ src/experiment_report.py
@@ -80,7 +80,7 @@ class ExperimentReport:
         preds = []
         for img_name, img_df in dataset.items():
             result = {
-                "boxes": torch.from_numpy(np.array(img_df["box"])),
+                "boxes": torch.from_numpy(np.array(img_df["box"], dtype=np.float64)),
                 "labels": torch.from_numpy(img_df["class_id"].to_numpy()),
             }
             if is_pred:
